Refactoring Reflection API
