+++ test/cyclic/src/cyclic.ts
@@ -1,6 +1,6 @@
-//import "./cyclic/main";
+import {Main} from "./cyclic/main";
 //import "./decor/decorated";
-import "./decor/decorated2";
+//import "./decor/decorated2";
 
 //console.info("Testing cyclic references");
-//console.info(main);
+console.info(new Main("config"));
