+++ test/cyclic/src/cyclic/four.ts
@@ -1,0 +1,3 @@
+export class Four {
+    constructor(){}
+}
