+++ test/cyclic/src/cyclic/one.ts
@@ -1,8 +1,10 @@
 import {Main} from './main';
+import {Three} from './three';
 
 export class One {
     constructor(){
         Main.from('One');
+        console.info('One',new Three());
     }
     private value:number = 56;
     public get length():number{
