+++ test/cyclic/src/cyclic/three.ts
@@ -1,0 +1,7 @@
+import {Four} from "./four";
+
+export class Three {
+    constructor(){
+        console.info(Four)
+    }
+}
