+++ test/cyclic/src/cyclic/two.ts
@@ -1,9 +1,10 @@
 import {Main} from './main';
-import {Meta} from "../decor/decors";
+import {Three} from './three';
 
 export class Two {
     constructor(param?:String){
         Main.from('Two');
+        console.info('Two',new Three());
     }
     public calc(param:String):Function {
         return Main.from.bind(this);
